Replace EditLinkTool with ExternalLinkEditor.
