+++ src/article/metadata/MetadataPackage.js
@@ -1,4 +1,4 @@
-import { AnnotationCommand, EditAnnotationCommand, getKeyForPath } from 'substance'
+import { AnnotationCommand, getKeyForPath } from 'substance'
 import {
   BasePackage, EditorBasePackage, ModelComponentPackage, FindAndReplacePackage
 } from '../../kit'
@@ -21,7 +21,6 @@ import {
   ReplaceFigurePanelImageCommand, RemoveFigurePanelCommand, OpenFigurePanelImageCommand
 } from '../shared/FigurePanelCommands'
 import EditEntityCommand from '../shared/EditEntityCommand'
-import EditExtLinkTool from '../shared/EditExtLinkTool'
 import EditInlineFormulaCommand from '../shared/EditInlineFormulaCommand'
 import EditInlineFormulaTool from '../shared/EditInlineFormulaTool'
 import EditXrefCommand from '../shared/EditXrefCommand'
@@ -143,10 +142,6 @@ export default {
     config.addCommand('edit-author', EditEntityCommand, {
       selectionType: 'author',
       commandGroup: 'author'
-    })
-    config.addCommand('edit-external-link', EditAnnotationCommand, {
-      nodeType: 'external-link',
-      commandGroup: 'prompt'
     })
     config.addCommand('edit-formula', EditInlineFormulaCommand, {
       nodeType: 'inline-formula',
@@ -327,7 +322,6 @@ export default {
 
     // Tools
     config.addComponent('add-figure-panel', InsertFigurePanelTool)
-    config.addComponent('edit-external-link', EditExtLinkTool)
     config.addComponent('edit-xref', EditXrefTool)
     config.addComponent('edit-formula', EditInlineFormulaTool)
     config.addComponent('insert-inline-graphic', InsertInlineGraphicTool)
