+++ src/article/shared/styles/_index.css
@@ -2,7 +2,6 @@
 @import './_card.css';
 @import './_checkbox-input.css';
 @import './_custom-metadata-field.css';
-@import './_external-link-editor.css';
 @import './_file-upload.css';
 @import './_form-row.css';
 @import './_heading.css';
